+++ src/oldnews/data/models/__init__.py
@@ -5,7 +5,11 @@
 from .local_article import LocalArticle, LocalArticleAlternate, LocalArticleCategory
 from .local_folder import LocalFolder
 from .local_state import LastGrabbed, NavigationState
-from .local_subscription import LocalSubscription, LocalSubscriptionCategory
+from .local_subscription import (
+    LocalSubscription,
+    LocalSubscriptionCategory,
+    LocalSubscriptionGrabFilter,
+)
 
 ##############################################################################
 # Exports.
@@ -17,6 +21,7 @@ __all__ = [
     "LocalFolder",
     "LocalSubscription",
     "LocalSubscriptionCategory",
+    "LocalSubscriptionGrabFilter",
     "NavigationState",
 ]